+++ src/main/java/school/sptech/crud_proj_v1/projection/PicoDiaProjection.java
@@ -1,6 +1,8 @@
 package school.sptech.crud_proj_v1.projection;
 
+import java.math.BigDecimal;
+
 public interface PicoDiaProjection {
     String getDiaSemana();
-    Double getFaturamento();
+    BigDecimal getFaturamento();
 }
