+++ src/main/java/school/sptech/crud_proj_v1/projection/ProdutoRentavelProjection.java
@@ -1,7 +1,9 @@
 package school.sptech.crud_proj_v1.projection;
 
+import java.math.BigDecimal;
+
 public interface ProdutoRentavelProjection {
     String getNome();
     Integer getVendas();
-    Double getLucro();
+    BigDecimal getLucro();
 }
